feat: provide `balance` command
The node manager makes it easy to list the rewards balances for all the nodes it's managing. It was
also requested that `status --details` show the balance.

There was also a request to include the store cost for the node, but this information isn't as
easily available. It may require an RPC extension to obtain it.